+++ R/outdated.R
@@ -56,7 +56,7 @@ outdated <-  function(
 ){
   force(envir)
   if (is.null(config)){
-    config <- config(
+    config <- drake_config(
       plan = plan,
       targets = targets,
       envir = envir,
@@ -166,7 +166,7 @@ missed <- function(
 ){
   force(envir)
   if (is.null(config)){
-    config <- config(
+    config <- drake_config(
       plan = plan,
       targets = targets,
       envir = envir,