+++ R/predict_runtime.R
@@ -248,7 +248,7 @@ rate_limiting_times <- function(
 ){
   force(envir)
   if (is.null(config)){
-    config <- config(
+    config <- drake_config(
       plan = plan,
       targets = targets,
       envir = envir,