+++ tests/testthat/test-Makefile.R
@@ -6,13 +6,13 @@ test_with_dir("recipe commands", {
   my_plan <- workplan(y = 1)
   expect_true(is.character(default_recipe_command()))
   expect_true(is.character(r_recipe_wildcard()))
-  con1 <- config(my_plan, command = default_Makefile_command(),
+  con1 <- drake_config(my_plan, command = default_Makefile_command(),
     parallelism = "Makefile", recipe_command = "some_command",
     verbose = FALSE
   )
   expect_equal(con1$recipe_command, "some_command")
   expect_true(con1$recipe_command != default_recipe_command())
-  con2 <- config(plan = my_plan, parallelism = "Makefile",
+  con2 <- drake_config(plan = my_plan, parallelism = "Makefile",
     recipe_command = "my_command", verbose = FALSE)
   expect_equal(con2$recipe_command, "my_command")
   expect_true(con2$recipe_command != default_recipe_command())
@@ -22,7 +22,7 @@ test_with_dir("no Makefile for make_imports()", {
   expect_equal(cached(), character(0))
   x <- workplan(a = ls())
   expect_false(file.exists("Makefile"))
-  con <- config(
+  con <- drake_config(
     x,
     parallelism = "Makefile",
     verbose = FALSE
@@ -36,7 +36,7 @@ test_with_dir("prepend arg works", {
   config <- dbug()
   config$verbose <- FALSE
   config$prepend <- "# add"
-  store_config(config = config)
+  store_drake_config(config = config)
   run_Makefile(config, run = FALSE)
   lines <- readLines("Makefile")
   expect_true(grepl("# add", lines[1]))
@@ -50,11 +50,11 @@ test_with_dir("files inside directories can be timestamped", {
   )
   plan$target[1] <- file <- eply::quotes(file.path("t1",
     "t2"), single = TRUE)
-  config <- config(plan = plan, targets = plan$target[1],
+  config <- drake_config(plan = plan, targets = plan$target[1],
     parallelism = "parLapply", verbose = FALSE,
     envir = new.env(), cache = NULL)
   path <- cache_path(config$cache)
-  store_config(config = config)
+  store_drake_config(config = config)
   run_Makefile(config, run = FALSE)
   expect_silent(mk(config$plan$target[1], cache_path = path))
   expect_true(file.exists("t1"))
@@ -78,7 +78,7 @@ test_with_dir("basic Makefile stuff works", {
   initialize_session(config = config)
   log_attempts(targets = outdated(config = config), config = config)
   config$recipe_command <- "Rscript -e"
-  store_config(config = config)
+  store_drake_config(config = config)
   run_Makefile(config, run = FALSE, debug = TRUE)
   using_global <- identical(config$envir, globalenv())
   if (using_global) {
@@ -112,7 +112,7 @@ test_with_dir("basic Makefile stuff works", {
   mk(targ, cache_path = cache_path) # Verify behavior when target is current
   expect_equal(unname(progress(list = targ)), "not built or imported")
 
-  store_config(config = config)
+  store_drake_config(config = config)
   run_Makefile(config, run = FALSE)
   expect_false(file.exists(globalenv_file(cache_path)))
 })
@@ -125,7 +125,7 @@ test_with_dir("Makefile stuff in globalenv()", {
     envir = globalenv(),
     verbose = FALSE
   )
-  store_config(drake_TESTGLOBAL_config)
+  store_drake_config(drake_TESTGLOBAL_config)
   run_Makefile(drake_TESTGLOBAL_config, run = FALSE, debug = TRUE)
   clean(list = targ)
   drake_TESTGLOBAL_config$cache$del(key = targ, namespace = "progress")