+++ NEWS.md
@@ -17,10 +17,12 @@
     - `backend()` => `future::plan()`
     - `build_graph()` => `build_drake_graph()`
     - `check()` => `check_plan()`
+    - `config()` => `drake_config()`
     - `evaluate()` => `evaluate_plan()`
     - `expand()` => `expand_plan()`
     - `gather()` => `gather_plan()`
     - `plot_graph()` => `vis_drake_graph()`
+    - `read_config()` => `read_drake_config()`
     - `render_graph()` => `render_drake_graph()`
     - `session()` => `drake_session()`
     - `summaries()` => `plan_summaries()`